refactor(spec-546): send 12 single-consumer symbols home to their handler (t-2)
std-51: a symbol with exactly one consumer belongs in that consumer. Twelve
of them move out of agent/handlers/shared.ts into the sibling that was
already their only caller, dropping their export on arrival — comments.ts x3,
tasks.ts x3, standards.ts x3, plus acs.ts, roles.ts, lifecycle.ts.

shared.ts: 1810 -> 1531 lines, 59 -> 24 exports. Nothing outside
agent/handlers/ changes its import specifier; the 56 importers are untouched.

Three corrections the compiler found in the plan, which said thirteen:

- handoffInterpolationContext does NOT move. shared.ts calls it itself at
  :949, inside composeGuidanceEnvelope, which stays. Two consumers, so the
  single-consumer rule does not apply — the measurement behind the plan
  counted consumers in OTHER files, which is the right test for "exported
  without a consumer" and the wrong one for "can it leave". docs.ts is
  therefore not touched at all, and the export count lands at 24, not 23.
- COMMENT_SNIPPET_LEN travels with formatTerseComment, its only user, as a
  private const. formatTerseComment's three internal callers all live inside
  formatDocCommentsTerse, which moves with it.
- parseTypeFilter was dead code, not merely a superfluous export: zero
  consumers, and routes/comments.ts:84 carries its own private implementation
  of the same name. Left private-and-unused it would be strictly worse than
  gone, so it is deleted, along with the ../../types/roles.js import that
  became wholly dead.

Five symbols that look like they belong here deliberately stay (dec-2, ac-2):
renderFooterSignal, craftActivityBlock, composeStatusOverview, StatusFacts
and relatedIssuesNudge each have one consumer too, but it is the tool-specs.ts
façade. Moving them there takes them outside the only directory
guidance-authoring-confined scans, and its `while (idx !== -1)` loop then
never runs its body: the test would pass while asserting nothing and the
"one author of footer prose" invariant would go unguarded, in green. The
façade's leading comment now names those six and explains the trap.

docs.ts's two unused imports (listDecisions, listTasks) are pre-existing —
confirmed by re-running the check with these changes stashed — and are left
alone [per std-51 cl-26: existing modules are not refactored on sight].

Verified: make typecheck exit 0; no unused import or local in any of the
seven touched files; full server suite 707 files / 6552 tests passed, same
counts as before the change; the seven source-scanning guards run explicitly,
45 tests passed, none modified; git diff --cached confirms the whole of PR1
stays inside agent/handlers/ plus a comment-only edit to tool-specs.ts, with
nothing under any *.test.ts.
